Extended contact search implementation
The original implementation only searched for the given name. Now there are
multiple input fields for given name, last name and company.

Documentation for the XML schema "directorySearch" doesn't seem to exists.
The search could be much improved if we would know how to OR attribute searches
and do pattern/regex matches for attribute values.