Updates to enable compilation in a Cygwin/MinGW environment
Makefile.mingw: updated the pidgin-source directory to a more recent version
src/Makefile.mingw: Removed zlib from libraries because it caused crashes for me
src/sip-ntlm.c: Removed zlib import and copied the crc32 function into code similar to how it works in gg and qq protocols in pidgin source.  Calling Zlib's crc32 caused a crash for me on windows.  Iconv calls were causing crashes on windows also, turned out they needed the '-' between 'UTF' and '8' or '16LE'.  Hopefully this won't cause any problems on *nix, but it should be tested, everything should work as long as 'UTF-16LE' and 'UTF-8' are listed in the output of 'iconv --list'.
src/sipe.c: Moved the #define _ into *nix specific part of the includes, because in WIN32 we include pidgin's "internal.h" which defines _ already.  Changed random() to g_random_int() because in WIN32 random() isn't defined, and it appears from other code that g_random_int() is a good way to generate pseudo-random numbers in pidgin.
src/sipe.h: Added #ifndef _WIN32 around the #define for PURPLE_WEBSITE because it is defined in "internal.h"
src/uuid.c: Duplicated the #ifndef _WIN32 around *nix headers like we do in sipe.c and sip-ntlm.c so that we wouldn't use any headers that aren't provided under Cygwin.  Similarly added an #ifdef around the code to retrieve the MAC address, since that doesn't work on Windows, so it has a hardcoded MAC for now.  This will need to be fixed for windows eventually.