Fixed the subscription process to use the new schedule system in OCS2007
To correct implementation for batched and single category Subscribe methods
based on [MS-PRES].

Now is sending a batched category Subscribe request for each contact in
the roaming contacts list. If some contact backs a "resubscribe" state then
is sending a single category subscribe request.

If the subscription-state is active the server is sending (BE)NOTIFY messages
with the current status for each contact online.

Also was implemented the setSubscribers answer. This answer is sending when
somebody subscribe us. The roaming-self Subscribe sets a dialog with
the server requesting for some change neither in categories, container
or subscribers. The server sends a (BE)NOTIFY messages if some data changes.
In this case is sending a setSubscriber anwser if is found some new subscriber
inside the roaming-data body.