Full support for Batched Category Subscription for OCS2007. The changes doesn't affect to LCS2005 where there isn't such support.
The support neccesary to have Batched subscription is that server backs
something like that: Supported: msrtc-event-categories

Then the variable msrtc_event_categories is set to a true value.
If we don't have that support we use Single Subscription how so far )in LCS2005 is the normal form).

The batched category subscription is faster reduce the number of request to the server on login.

All seems works fine now on OCS2007.