Fixed in the vnd-microsoft-roaming-self+xml subscription request. The first subscription is a request for initial session with the server. The server gives us a full information about: categories, container, subscribers.
The subsequent (BE)NOTIFY
messages sending by the server are updates. The problem is that sent the same
Content-Type header (vnd-microsoft-roaming-self+xml).
Then we need to try different from the subscription request.

On LCS2007 if we don't do that we fault in an infinite loop. Because the first
subscription give us a full subscribers list, this is parsing and re-send with
an INVITE request that the server answers with another (BE)NOTIFY and so ...
forever.