Fixed the resubscription process when expires comes to 0. On OCS2007 and LCS2005 the server is sending continuously messages (BENOTIFY, NOTIFY or 200 ok) to update the status for each contact subscribed previously.
Each message contains the header
  subscription-state: active;expires=time

where the time is decresing between 200 to 600 segs each new update.
We need to resubscription before te time comes to 0 or the subscriptions
is terminated (the server doesn't send more updates).

Also, was added some extra headers on the INVITE request, trying to be
compatible with the CHOWA client and others.