Fixed the problem when for some contacts the server sends back a "terminated" status (described on the previous commit).
The problem is that some users are from the same domain but
is hosted on a different server or server pool. In this case, the server
sends back a poolFqdn attribute, which supplies a URI that routes
the request to a server or server pool that can accept a batch subscription for the user.

[MS-PRES] Section 3.4.5.1.3 Processing Details

I reverted the previous commits because the contacts should be subscribed now.