Fixed crash during server redirect
With the new security framework the gssapi_context was not cleared when the
authentication data structure was released. Therefore the signature code was
assuming the context was still valid and tried to use the other authentication
data which was already set to NULL.

To fix this I needed to add a context destructor to the framework. While doing
this it became very clear that the API and code could be simplified a lot by
adding the module functions to the context.

Please note that I only updated the NTLM module to the new context API as
I can't compile the Kerberos code on my system. You *WILL* get compilation
errors if you try :-)