Send presence subscribe requests to poolFqdn servers
If the server requests a presence resubscription and supplies a poolFqdn
attribute then we'll now send a batched request for all user with the same
poolFqdn value. We also get only one response to this request so when we set up
a scheduled action for the resubscription we need to look into the response
for the URIs and can't use To:/From:.

Although this is the same behaviour as the M$ OCS client implements it still
doesn't solve the "subscription terminated" problem. We now get the termination
not from the user but only one from the home server for multiple users.

Scheduled actions can now supply a destroy function for the payload.

Found another memory leak in the new security framework.