Clean up connection configuration and setup
This is a partial fix for bug #2833015.

Now there are only two connection settings: server[:port] and type. If the user
leaves the server setting empty then server auto-discovery via DNS SRV records
is used. The old code did also not handle the case of user-specified server and
type == "auto".