Make password mandatory again.
NTLM requires a password. Don't crash if none is provided.

When the protocol option "password is optional" is set then the only way to
specify the password is via the saved password in the account settings. But
those passwords are stored unencrypted in libpurple's configuration files.
Not very secure :-(

There seems to be no way for a protocol plugin to initiate a password dialog
when it has decided that it needs one after all. For now we have to live with
the fact that the password has to be mandatory...