Fix timeout calculation in sipe_buddy_subscribe_cb()
If RAND_MAX == INT_MAX then the calculation for "timeout" always returns 0,
because the dividend can never be bigger than the divisor as it would then
need to exceed maximum integer range.

Convert the calculation to use unsigned integers. Multiply both terms of the
formula with 1 / time_range to scale the calculation to the limits of UINT_MAX.
It now should return a value in the right millisecond range.