Fix memory leaks in sipe_buddy_menu()
valgrind doesn't show the the g_strdup()'ed callback data as memory leaks for
some unknown reason. But logic dictates that you can't allocate memory for the
data, because who is going to free it when the menu is destroyed again without
calling any of the callback functions???

Also some DRY cleanup for the buddy property name constants.