Fix for bug #2859241: conference invite destroys IM dialog
The original code sent a BYE using the conference invite dialog, i.e. using the
wrong Call-ID, after receiving the OK for our conference invite. Then we
received the BYE for the conference invite from the remote party and deleted
our IM dialog. The next IM message to the same remote party creates a new IM
dialog on our side but the INVITE is rejected by the server as it still has an
active IM dialog from the previous IM.

If we have an active IM dialog with a remote party and invite that party to a
conference then we should send out a BYE for THAT dialog after we have received
the OK to the invite.