Fix: statuses stopped updated after ~8 hours. very ugly bug: Symptoms: After about 8 hours (usually noticed on the next working day) contacts from your Home server stopped to update their statuses.
Problem in code: SELF-contact was added to Batch list for subscription to
External pool (not Home one). Not initially, but on resubscription
after timeout (hence ~8 hours delay).

How bug worked: Prerequisite - some contacts reside on external pool.
When we resubscribe to external-pool's contacts, on the response to
that we gather contacts from that response and scheduled resubscription
after the timeot. Problem was that we added SELF to that list due to
not very careful XML parsing. As a result of the iteration after the
timeout fired - we send subscribe with self to External pool, got
pointed to our home server for subscription. Then our logic sent
subscription to our Home server with ONLY one SELF contact. That terminated
previously batched subscription for ALL the contacts on our Home server.
So their statuses effectively stopped updated.