conference: rejoining after reconnect by typing message in the conversation window left open.
The point is to join to accidently disconnected conversation with your team
for example due to network outage (VPN fault). Before that it
required to ask one of chat participant to invite you back to
the chat. Very annoying if happens several times a day.
Not without a hack. Purple mark chat conversations as left and doesn't alow
to sent messages from there (no callbacks to plugin). It took
to mark all remaining purple chats as NOT left on (re)startup to let user
type there and send. But after that in was needed to mark chat back to left state to allow purple to rejoin chat to the SAME left window.