fix 2882658 - SIGSEGV in process_incoming_info_csta() My theory on the issue is that your CSTA Gateway - Cisco-Systems-UA\7.0 (interestingly, we've tested against exact same gateway) - was confused with the phone number provided:
tel:5555551234v:41234

That's why your phone goes off hook but doesn't dial the number.
Where is 'v:' came from in the tel string? It should be something like
tel:5555551234;ext=41234 to my mind to be a valid tel URI.

The crash itself goes of the fact that CSTA returns with empty XML body:

Content-Type: application/csta+xml
Content-Length: 0

I don't see much point in this empty message, but this is how it worked in
your case.