Take PURPLE_INIT_PLUGIN into use
Since the beginning there has been a copy of the code generated by this macro
at the end of sipe.c. At least the Adium developer has run into integration
problems because of this.

I never understood the cryptic comment associated with the old code until I
changed to the original macro and noticed we were defining PIDGIN_PLUGINS
instead of PURPLE_PLUGINS and we were defining it at the wrong place.

I hope this doesn't break the Windows build...