openSuSE Build Service spec builds Debian 5, xUbuntu 8.04(LTS) - 9.10 This is for OBS, not Debian nor Ubuntu distributions.
The spec allows to use same original source distribution
for both rpm and deb packaging.

DON'T REMOVE IT
or provide precise URL to official document that prohibit it.

And nothing prevents to not include this folder to
source release if desired. It's located in contrib/ dedicated folder,
not even in root folder.

Point to keep spec in version control is to involve
as many interested parties as possible to improve the
packaging, fix errors and incorrectnesses.

I believe official packaging for Debian should follow
this same open approach, as it's even now has a lots to
improve, but not to hide.

If we had this set up before 1.8 release, the codebase
could be tested, thus reveal backward compatibility
problem on early stage.