ft: cancelling FT if rejected by server Not completely happy with outcome - Cancelling the FT appeared like it was cancelled by remote party, not server:
"(4:14:28 PM) Offering to send [MS-SIP].pdf to "Felix Fox"
(4:14:28 PM) "Felix Fox" cancelled the transfer of [MS-SIP].pdf
(4:14:28 PM) Your message or invitation was not delivered, possibly because it contains a hyperlink or other content that the system administrator has blocked."

and similar Error window.

But definitely better than leave FT in in-progress state.