NTLM: MAC compare string needs to use upper case letters
MAC() uses sprintf("%02X").

Added two more MAC tests. Interestingly enough the random pad is completely
useless, as its encrypted value is simply overwritten. This seems to be a
design flaw in NTLMv1 and I guess that's why the original code only compared
the last 16 characters of the NTLM signature...