core cleanup: implement GMime-based sipe_mime module
If GMime 2.4 or 2.6 headers are available then the Purple MIME implementation
is disabled and replaced with the one based on GMime. If another backend than
purple is enabled then GMime is a mandatory build requirement.

The minimum versions are 2.4.16 or 2.5.2 in order to get this bug fix:

    <https://bugzilla.gnome.org/show_bug.cgi?id=613653#c8>

It also available in version 2.2.26 but GMime upstream maintainers request that
we shouldn't use this old obsolete API. That means that on platforms that only
offer the GMime 2.0 API you'll only be able to compile the purple plugin. It is
doubtful if such outdated platforms would have version 2.2.26 available or
would support other frameworks, like telepathy, so this is no real limitation.