xml: fix overflow for integer attributes >= 2^31
atoi() and friends return rubbish on 32-bit(?) platforms for integers >= 2^31.
Change API to return guint and use g_ascii_strtoull(), which returns a 64-bit
signed integer. That makes sure that integer values with up to 32 bits are
handled correctly.

This is the root cause for the problem addressed in commit

   0e7f5dfe8531b1dfc1397af9e0ea1c5500a1e161