build: don't use _LDFLAGS for -L/-l options
_LDFLAGS contents appear before _LIBADD/_LDADD contents on the libtool comand
line. Therefore -L/-l options in _LDFLAGS are not in the correct order.

Usually this doesn't matter but in OBS openSUSE 11.2 repository the linker
option --as-needed is enabled by default. This option demands the correct
order of objects, linker libraries and shared libraries in order to resolve all
references.

Moved all _LDFLAGS contents to _LIBADD/_LDADD.