core cleanup: introduce new sipe core data structure
Start of phase 3 of the core cleanup.

sipe_core_public contains the fields accessible to the core and the backend.
sipe_core_private contains the fields only accessible to the core.
sipe_core_private starts with sipe_core_public and therefore pointers to them
are interchangeable.

The purple backend stores a pointer to sipe_core_public in gc->proto_data.