build: add NSS auto-detection
If NSS development files are available then include sipe-crypt.c instead of
purple-crypt.c in the build. Add NSS as build requirement for non-purple
backends.

NOTE: as sipe-digest.c is not purple-free it is not yet included.

Updated packaging files with nss development package information verified with
OBS builds. The package is currently commented out.