transport: always send an empty line after a message
This was one of the old code lines not touched during the rewrite. When a
message was submitted to the transport layer the code looked if we were already
sending something, i.e. if something was in the transmit buffer. In that case
it first appended a line terminator to the buffer to make sure that the
previous message is correctly line terminated, before adding the new packet.
Frankly I don't understand the logic behind that...

Now sendout_pkt() always makes sure that a message ends with an empty line,
i.e. it appends two line terminators to the transmit buffer. I have now
started pidgin tens of times and never got an "Invalid Signature" error as
reported in bug #2834758. It might be that we finally identified the root
cause for that problem. I keep my fingers crossed...

I'm also wondering if this might be the root cause for the problem that Sipe
has to insert artificial pauses when it has to send a LOT of messages, e.g.
subscription requests at startup. These were introduced "as not to overload
the server" but I think we should revisit the logic behind that after the
changes included in this commit.

Unfortunately not all problems are solved. In my tests I received at least
twice an "Authentication Failed" error. But I'm now inclined to think that
root cause for that is a different one as the one for bug #2834758.

I'm also not sure if the "fast typing causes invalid signature" error as
discussed in this thread

  <http://sourceforge.net/projects/sipe/forums/forum/688535/topic/3671326/>

is fixed by this commit. I'm having a hard time to reproduce it reliably. As
the received message starts with the last characters of the previous message
we sent to the server, I'm leaning towards that this commit is still not
enough to make sure that the server can reliably detect the boundary between
two messages we sent to it. I'm wondering if something like a flush operation
is required between two messages.