transport: when connect to DNS SRV result fails try next one
Re-implemented the old current_service in a better way.

While we are trying to resolve DNS SRV records sipe_private->service_data is
non-NULL. If a DNS SRV record was resolved successfully but the connection
attempt failed then we check service_data and simply call resolve_next_service,
which will initiate a resolve attempt for the next service in the list.

service_data is set to NULL on a successful connection attempt, when a
connection is closed or the last service in the list has been tried.