mime: update GMime implementation to latest API
Valgrind also showed memory leaks in the GMime implementation. I added a
g_object_unref() for the multipart we construct. Unfortunately the GMime
documentation is very unclear about memory handling :-(

This change fixes the memory leak. But now for each part in the multipart a
g_object_unref assert message is generated. My guess is that this is a bug
in GMime...