Fix broken sipe_ht_equals_nick()
In commit 99cd5c6ab1c4fe3007d3a4603dadad905359ae75 I replaced purple util.h
functions for the core cleanup. Unfortunately there was a typo in the hash
table string compare function which caused only the second string to be
compared to itself, i.e. the function always returned TRUE. This of course
causes havoc for the hash table functionality.

The broken hash table functionality was noticed by a RHEL5 user. On that
platform the hash bucket size seems to be smaller and thus the broken
broken behaviour more likely.

The broken code has been in the 1.10.0 release. As it affects the buddy list
it could be the root cause for some of the "buddy appears offline" reports.