Remove dialog when response to IM message is 408/480/481
This is a continuation of commit fa71e0bc0bf378f950a9f78fed8e780a07b62c79.

In the case that our IM message was rejected by the server he also can reject
our BYE, i.e. we never receive a proper response to it. Therefore we will not
close our side of the dangling IM session. The next messages will then fail
because of the same problem until the session is timed out on the OCS server
after 10 minutes.

The code now removes the dialog after sending out the BYE so our next attempt
will trigger an INVITE.