xml: simplify entity decoding for attribute values
Further studying of libxml2 documentation and some testing revealed that we
don't need a general purpose entity replacement routine after all. All entities
are decoded except "&amp;". This is replaced by the equivalent "&#38;" which
we can handle with a simple replace() call.

Tested by creating a SIPE group with the name "Test & < > " ' Test" which is
handled correctly.