purple: add callback to generate data for add chat dialog
Unfortunately the chat_info_defaults() callback only gets the chat name and not
the full room infomation from the room list. Therefore we have to keep a map
name -> URI so that we can fill in the 'uri' field.

It would be nice to fill the "Alias" entry with the room name. But I couldn't
find a way how to do that...