groupchat: first try to send message to channel
Doesn't seem to work though, at least on my test server. The log shows an error
message after a timeout:

SIP/2.0 503 Service Unavailable
Ms-diagnostics: 20000;source="xxx";reason="Chat service busy";component="GroupChat"

This could either be a real server problem or that we don't follow the correct
protocol.

Compared to the example M$ Group Chat logs we skip some messages after we join
a room and we leave out some information from the chatgrp message. But without
XCCOS protocol spec it is difficult to say if they are relevant or not :-/