groupchat: implement support for ntc:YYY messages
"Look Ma' what I discovered in the log..."

Yup, testing with two accounts on the same server revealed a completely type
of messages, e.g. ntc:join and ntc:bjoin. The server sends those to inform us
that a new user has entered the channel.

The structure is the same as for rpl:XXX. So we can use the same processing
function.