Fix for bug #3082602: Crash on Autodiscover
When we get a SSL error we should close the gsc immediately. It seems that
there are situations when purple gives us an SSL error immediately after open
but then continues with the SSL handshake. The handshake then triggers a
callback with already freed data.

Well, that is at least what I think might be going on...