Fix for bug #3082602: Crash on Autodiscover
The real problem was that we received a redirect response with an invalid URL.
But the check for that was at the wrong place and let to a use-after-free
situation.

Now we parse the URL first and if it is valid then we do the cloning of the
http_conn for the redirect.