Fix #3090663: Re-authentication fails (2nd attempt)
Commit f41ea03faed867cb7261b1a7f99a2367d2075919 had the right idea but there
were two problems with it:

 - wrong authentication data, e.g. wrong password, was no longer detected and
   caused an endless re-authentication loop.

 - the pending transaction of the rejected REGISTER message was not removed,
   causing dangling transactions to pile up over time.

While trying to figure out how to handle the first problem correctly, I noticed
that the information in transport->register_status was useless. I removed the
field and all references to it from the code.

I also cleaned up some of the debug messages in sip-transport.c.