Fix #3090663: Re-authentication fails (4th attempt)
Commit b8ecabd1ba8161a0b30067d478a993866778058e changed the code to drop a
message and its associated transaction if it was unsigned and not a REGISTER
or 401 response.

It turns out that OCS can send provisional messages (100 Trying) that aren't
signed. The old code simply ignored them but with the new code the transaction
got dropped. As a consequence no IM session could be established, because when
we received the final 200 response to our INVITE the transaction was gone.

Change the code to only remove transactions for unsigned messages that have a
response code of 200 or higher.
(cherry picked from commit 704bc7a923250d5ce87fc8b9308fd806a70ba07b)