Handle 401 response to non-REGISTER requests correctly
After we have started a full reauthentication any non-REGISTER request will be
rejected by the server with 401 Unauthorized.

The old code disrupted the authentication handshake, because it messed with the
authentication data, and therefore caused an incorrect authentication failure.

The new code just resends the original non-REGISTER request. This creates a
request - 401 response loop. Once the authentication handshake is completed we
break the loop by re-signing the request which is then accepted by the server.