conference: Allow others to initiate voice conference if sipe supports it
When we build sipe with voice and video support, send "audio-video"
entity in addConference requests. Otherwise other participants will not
be able to add voice to the conference (buttons in Communicator are
rendered as disabled).

In sipe builds without V&V I left the request unchanged as it is a good
thing, when we can do it,  to disable features that we do not support.

All of this applies only if we are the creator of conference.