Fix "undelivered message" errors when client reconnects after crash
When Sipe reconnects after a crash, we are not able to send messages
to contacts with which we had open conversations when the crash
occured. Server sends error response with reason="This client has an
IM session with the same conversation ID"

Setting random Ms-Conversation-ID prevents this problem so we can
continue the conversation.

For now we don't have other usage for the value, so it is freed
immediately after we create an INVITE.