Fix #3156430: Messages not Delivered
When SIPE received a new INVITE for an existing session & dialog, it only
updated the call ID for the session but not the dialog. The next user messages
where therefore still sent with the old call ID and promptly rejected by OCS
with a 400 "malformed route header" error.

Update the call ID in the dialog too.