Fix sipe_buddy_menu_copy_to_cb when copying buddy not yet in our list
When having IM session with someone not in our buddy list, right click
on header of the conversation pane and selecting Copy To>(Group Name)
adds the contact to list, but is permanently shown as offline and is
erased again when client restarts.

Call sipe_add_buddy instead of sipe_group_buddy. sipe_buddy is then
created when needed, we subscribe to its presence and add it to
contact list on server.