Fix #3156430: Messages not Delivered (3rd attempt)
From the provided M$ Communicator logs it became clear how it behaves: if it
receives an INVITE for an already existing dialog then it sends out a BYE for
the original one, drops it and then continues setting up a new dialog.

Updated process_incoming_invite() to do the same, thus getting rid of most of
the code for the special case. It no longer needs sipe_dialog_parse_routes(),
so that function was removed from the sipe-dialog.c API.

I was able to reproduce the situation once with the original code and once with
the new code. With the new code we no longer get an error from OCS.