Fix #3161273: Lost Connection Gives No Error Message (part IV)
Implement re-enqueue of unconfirmed messages when we receive an INVITE to an
already existing dialog.

If we receive an INVITE to an already existing dialog then all unconfirmed
messages have never arrived at the remote party. This can happen when the
remote party disconnects, then reconnects and we never received the BYE for
the dialog. We re-enqueue the messages so that they are sent out again after
the new dialog has been created.