Fix #3161273: Lost Connection Gives No Error Message (part IX)
Remove unprocessed messages when error response is received for INVITE.

When a user types a message for a session which currently doesn't have a dialog
then this will trigger an INVITE and the message is queued on the outgoing list
for sending it after the new dialog has been established. If the user keeps
on sending messages but the INVITEs fail then we tell him that the message was
not delivered but it is left on the outgoing list. The next time an INVITE
succeeds all those messages are sent out in one go.

Removing unprocessed messages after an error response was received for the
INVITE fixes this problem.