Fix #3161273: Lost Connection Gives No Error Message (part X)
Generate an error for each unprocessed message when INVITE fails.

If the response to the INVITE is delayed then the user has a chance to type
additional messages that are queued on the outgoing list for sending after the
new dialog has been established. But when we receive an error response to the
INVITE we only generate an error for the first message.

Repeat the error for all unprocessed messages.