media: support dynamic addition and removal of video during call
Office communicator supports this, updates in SDP messaging on the
side of Sipe was needed. To make this work, also some bugs in Farsight
were fixed and support of removal of farsight streams and session in
libpurple implemented. The latter patch is attached and works, but
still some improvements are needed before it is submitted to upstream.

Related fs2 bugs:
https://bugs.freedesktop.org/show_bug.cgi?id=34580
https://bugs.freedesktop.org/show_bug.cgi?id=34582