im: fix response/timeout for multiparty/conference messages
Chat messages for multiparty/conference sessions are sent using the IM message
handling, i.e. they are also added to the unconfirmed message queue. But in
process_message_response/timeout() we were only handling IM sessions, so those
messages were incorrectly left on the queue and therefore after the session
was closed we printed error messages in the chat.