Fix #3267073: False "could not be delivered" errors
The logs provided by the reporters indicate that OCS2005 clients DO NOT
acknowledge our SIP MESSAGEs. Therefore the message timeout is triggered
and messages are reported to the user as not delivered.

Either this is a bug in the OCS2005 client or we do something wrong in our
SIP MESSAGEs. This hack removes the message timeout and is provided for
users who need to communicate with a still existing OCS2005 user base.

Do not enable it by default!