filetransfer: do not process incoming invite in "ms-text-format" header
There was problem sending files between two SIPE instances when
Forefront Security is running on OCS, because it seems FS does not
understand ms-text-format header.

We should not place "Supported: ms-text-format" to the invitation
replies and force the sender to resend the invite once again in
plaintext. Otherwise when receiver accepted the transfer, Forefront sent
CANCEL to the sender and kept receiver waiting for the file forever.

This was affecting only the first attempt to make a filetransfer,
second try, when session is already established, was processed
correctly.

We still must place "ms-text-format" header in SIP INVITEs, as MSOC
uses it to display notification in toast. Without it communicator
incorrectly displays two separate toasts "invite to conversation" and
"incoming file transfer".