presence: correctly detect multiple points of presence
Mystery solved: while I was writing the original implementation I was online
with Exchange OWA 2010, which acts as an OC client. That explains the
additional device instance in my logs, doooohhhhh...

Now we count the *unique* instances of the category "device" listed in
roaming-self. If we detect more than one instance than the user is online with
multiple clients.