media: drop support for ms-early-media
Remote party can have multiple active points of presence where the user
can accept call. Because libnice does not support parallel ICE
negotiation with multiple endpoints having the same set of candidates,
codecs, usernames and passwords, I drop support of early media from SIPE.

Reason is that on the call initialization, ICE negotiation was started
with the endpoint whose 183 Session Progress response we received
first. Later if remote user accepted the call on different endpoint, the
connection could not be established.

Now ICE process is delayed until final reply is received from the user
selected endpoint, so that connection establishment can be done only
with the single correct point of presence.

Drawback is a media transport lag between accepting a call and a moment
when ICE process is finished, but still better than inability to establish
the network route for media at all.