mingw: sync SPEC changes to src/core/Makefile.mingw
This drops all references to sip-sec-ntlm.c and NSS if USE_SSPI is
defined. This is the default, i.e. the behaviour should be unchanged
from earlier builds, but the binary should be smaller.

By commenting out the variable the user can disable SSPI and fall
back to sip-sec-ntlm.c and NSS.

These changes are untested as I can't run the MinGW build.