OBS: update files
Added support for mingw64 build.

Enabled gstreamer-devel for openSUSE > 11.4. openSUSE 12.1 will be the first
release to support Voice & Video out-of-the-box.

Renamed DSC file from xUbuntu-9.10 to xUbuntu-11.10. Updated the build
requirements to match what is required for Voice & Video support. Ubuntu 11.10
will be the first release to support Voice & Video out-of-the-box.