transport: add support for incomplete authentication
When TLS-DSK doesn't have a valid user certificate it needs to request
one first from the Certificate Provisioning Service. We can't send any
messages while this is ongoing. Add a flag for this situation.

This should only happen during registration. So in the current state the
registration is aborted in the case of TLS-DSK after the initial
REGISTER request. Once the certificate fetching code is implemented it
will have to retrigger the registration.